requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
playwright==1.46.0
pydrive2==1.20.0
gspread==6.1.2
google-auth==2.34.0
reportlab==4.2.2
Pillow==10.4.0
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
scikit-image==0.24.0
